User login/Logout working
Login button changed to a type of toggle button. If the user is logged
in, the image source will point to the 'logout' image, and when clicked
the user will be logged out. Logged out, the source will point to the
'sign in' image, and when clicked will direct the user to the signin
page.

Confirmation added on the image upload on the registration page.