Profiles show posts dynamically
Fixed a few bugs:

* The IM banner on the bottom right hand corner prevented users from
clicking anywhere on the page at the same height level as the banner
(which was always at the bottom of the page, for instance you couldn't
click on buttons at the bottom of the page)
* Fixed the CSS where there is a huge blank opening at the bottom of
some pages
* Fixed some stored procedures to properly work with the BLL
* Added a stored procedure and method to the Post class to retrieve a
topic name on a specific post.
* Changed the profile posts to display in a table rather than divs, this
helps with the aligning of the post heading as well as topic heading.
* Added check blocks to see if the table returned actually has something
in it before trying to access its rows (Which will result in an error),
where it returns a null where it would have returned an error