Fixed Date Validation
The problem was that the culture (Time/region etc) on the machines was
different, so the format which the date validator worked with differed
from the computers internal format.

The culture of the website was set to 'en-US' (English - United States)
in the web.config file for the validation to work properly.